clean up dnf cache and merge RUN commands
I previously didn't clean up the cache, since distrobox installs stuff on
entering the container, which means it was going to download the cache anyway,
so I thought I should just include it in the OCI image. This has the advantage
that dnf doesn't have to make a bunch of requests to redownload it, it's
included when pulling the image. However, this meant the cache could be outdated
by up to 6 hours. Also, in the next Fedora release, we should get dnf5 which
will hopefully speed up the process of downloading metadata.

I also previously split the RUN commands in the Containerfile so that podman
could hopefully cache some previous layers, which is good for the slow internet
I have. However, the podman Github Action I use squashes the image with no way
to disable, so it didn't help at all, instead just adding ~1MB to the image.