toolbx: switch to using systemd instead of flatpak to spawn stuff on the host
add a rust tool that calls systemd-run. this replaces host-spawn, which uses
the flatpak session helper to spawn programs. is this faster? no, because
host-spawn doesn't spawn another binary, but this tool does. but i like systemd.
maybe i'll change this in the future to use the d-bus api directly, but i assume
that means having to do a bunch of pty stuff, which i don't want to do (sounds
hard).